Fix 'Take from Previous' to take the correct period
The option 'Take from Previous' did not work correctly if there are
entries in the future. In that case, the function
get_previous_period_info() returns the last entry (which is in the
future) and not the one right before the current one.
I changed the algorithm in get_previous_period_info() to return the most
recently ended period that is in the past.